+++ src/components/OrdersHistory/OrdersHistory.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const OrdersHistory = () => {
+    return(
+        <div>
+            OrdersHistory
+        </div>
+    )
+}
+
+export default OrdersHistory
